# web/openapi.json
{
  "components": {
    "schemas": {
...
      "DeskPositions": {
        "properties": {
          "as_of": {
            "format": "date",
            "title": "As Of",
            "type": "string"
          },
          "desk_code": {
            "title": "Desk Code",
            "type": "string"
          },
          "desk_name": {
            "title": "Desk Name",
            "type": "string"
          },
          "positions": {
            "items": {
              "$ref": "#/components/schemas/DeskPosition"
            },
            "title": "Positions",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          }
        },
        "required": [
          "as_of",
          "run_id",
          "desk_code",
          "desk_name",
          "positions"
        ],
        "title": "DeskPositions",
        "type": "object"
      },
...
      "FactorsLatest": {
        "properties": {
          "as_of": {
            "format": "date",
            "title": "As Of",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          },
          "ticks": {
            "items": {
              "$ref": "#/components/schemas/FactorTick"
            },
            "title": "Ticks",
            "type": "array"
          }
        },
        "required": [
          "as_of",
          "run_id",
          "ticks"
        ],
        "title": "FactorsLatest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlaSummary": {
        "properties": {
          "end": {
            "format": "date",
            "title": "End",
            "type": "string"
          },
          "ks": {
            "title": "Ks",
            "type": "number"
          },
          "n_obs": {
            "title": "N Obs",
            "type": "integer"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/PlaPoint"
            },
            "title": "Points",
            "type": "array"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          },
          "spearman": {
            "title": "Spearman",
            "type": "number"
          },
          "start": {
            "format": "date",
            "title": "Start",
            "type": "string"
          },
          "window": {
            "title": "Window",
            "type": "integer"
          },
          "zone": {
            "enum": [
              "GREEN",
              "AMBER",
              "RED"
            ],
            "title": "Zone",
            "type": "string"
          }
        },
        "required": [
          "scope",
          "window",
          "start",
          "end",
          "n_obs",
          "spearman",
          "ks",
          "zone",
          "points"
        ],
        "title": "PlaSummary",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "EOD market-risk platform for a mock three-desk book. VaR/ES reported as positive potential loss, USD; P&L signed.",
    "title": "RiskDesk API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/backtest/pla": {
      "get": {
        "description": "P&L attribution: Spearman and KS between daily hypothetical and\nrisk-theoretical P&L with MAR32-style zones. Shipped with the options\nsleeve, which gives the statistic structural content beyond the linear\nlegs' log-linearization.",
        "operationId": "backtest_pla_api_v1_backtest_pla_get",
        "parameters": [
          {
            "description": "'FIRM', a desk code, or 'desk:CODE'",
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "FIRM",
              "description": "'FIRM', a desk code, or 'desk:CODE'",
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "description": "Trailing P&L days",
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "default": 250,
              "description": "Trailing P&L days",
              "maximum": 1000,
              "minimum": 20,
              "title": "Window",
              "type": "integer"
            }
          },
          {
            "description": "Pin to a run date (default: latest completed batch)",
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Pin to a run date (default: latest completed batch)",
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Backtest Pla"
      }
    },
    "/api/v1/backtest/summary": {
      "get": {
        "description": "Kupiec, Christoffersen and the Basel traffic light over the trailing\nexception series (observed P&L vs the prior day's VaR).",
        "operationId": "backtest_summary_api_v1_backtest_summary_get",
        "parameters": [
          {
            "description": "'FIRM', a desk code, or 'desk:CODE'",
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "FIRM",
              "description": "'FIRM', a desk code, or 'desk:CODE'",
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "model",
            "required": false,
            "schema": {
              "default": "HS",
              "enum": [
                "HS",
                "FHS"
              ],
              "title": "Model",
              "type": "string"
            }
          },
          {
            "description": "Trailing P&L days",
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "default": 250,
              "description": "Trailing P&L days",
              "maximum": 1000,
              "minimum": 2,
              "title": "Window",
              "type": "integer"
            }
          },
          {
            "description": "Pin to a run date (default: latest completed batch)",
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Pin to a run date (default: latest completed batch)",
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Backtest Summary"
      }
    },
    "/api/v1/desks/{desk_code}/decomposition": {
      "get": {
        "description": "VaR-decomposition waterfall (factor-class buckets + diversification) and\nthe desk's factor exposures. Buckets are empty for runs without the position\nstep (backfill runs) - the desk VaR itself is still reported.",
        "operationId": "desk_decomposition_api_v1_desks__desk_code__decomposition_get",
        "parameters": [
          {
            "in": "path",
            "name": "desk_code",
            "required": true,
            "schema": {
              "title": "Desk Code",
              "type": "string"
            }
          },
          {
            "description": "Pin to a run date (default: latest completed batch)",
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Pin to a run date (default: latest completed batch)",
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeskDecomposition"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Desk Decomposition"
      }
    },
    "/api/v1/desks/{desk_code}/positions": {
      "get": {
        "description": "Per-position standalone/component/marginal VaR for one desk (no\npagination - the book is under 100 rows by construction).",
        "operationId": "desk_positions_api_v1_desks__desk_code__positions_get",
        "parameters": [
          {
            "in": "path",
            "name": "desk_code",
            "required": true,
            "schema": {
              "title": "Desk Code",
              "type": "string"
            }
          },
          {
            "description": "Pin to a run date (default: latest completed batch)",
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Pin to a run date (default: latest completed batch)",
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeskPositions"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Desk Positions"
      }
    },
    "/api/v1/factors/latest": {
      "get": {
        "description": "Latest level and day move per active factor - the dashboard's factor\ntape. Moves are reported in each factor's own convention units.",
        "operationId": "factors_latest_api_v1_factors_latest_get",
        "parameters": [
          {
            "description": "Pin to a run date (default: latest completed batch)",
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Pin to a run date (default: latest completed batch)",
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactorsLatest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Factors Latest"
      }
    },
...
    "/api/v1/risk/exposures": {
      "get": {
        "description": "Key-rate DV01s per desk from the run's bootstrapped par curve - empty\nrows for runs that skip the curve step (backfill runs do).",
        "operationId": "risk_exposures_api_v1_risk_exposures_get",
        "parameters": [
          {
            "description": "Pin to a run date (default: latest completed batch)",
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Pin to a run date (default: latest completed batch)",
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyRateExposures"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Risk Exposures"
      }
    },
    "/api/v1/risk/history": {
      "get": {
        "description": "1-day VaR/ES, P&L and exception flags per date for one scope - the\ntimeline and sparkline feed.",
        "operationId": "risk_history_api_v1_risk_history_get",
        "parameters": [
          {
            "description": "'FIRM', a desk code, or 'desk:CODE'",
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "FIRM",
              "description": "'FIRM', a desk code, or 'desk:CODE'",
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "description": "Trailing run dates",
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "default": 90,
              "description": "Trailing run dates",
              "maximum": 1000,
              "minimum": 1,
              "title": "Window",
              "type": "integer"
            }
          },
          {
            "description": "Pin to a run date (default: latest completed batch)",
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Pin to a run date (default: latest completed batch)",
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskHistory"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Risk History"
      }
    },
    "/api/v1/risk/movers": {
      "get": {
        "description": "Desk VaR moves vs the run `window` days back, with the as-of day's\nlargest factor moves as driver strings - the Overview movers table.\nEmpty rows on the first run.",
        "operationId": "risk_movers_api_v1_risk_movers_get",
        "parameters": [
          {
            "description": "Pin to a run date (default: latest completed batch)",
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Pin to a run date (default: latest completed batch)",
              "title": "As Of"
            }
          },
          {
            "description": "Calendar days back for the delta baseline (drivers always describe the as-of day)",
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Calendar days back for the delta baseline (drivers always describe the as-of day)",
              "maximum": 30,
              "minimum": 1,
              "title": "Window",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskMovers"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Risk Movers"
      }
    },
    "/api/v1/risk/summary": {
      "get": {
        "description": "Firm and per-desk VaR/ES with limits, utilization, day-over-day moves\nand the diversification benefit - the dashboard tiles in one call.",
        "operationId": "risk_summary_api_v1_risk_summary_get",
        "parameters": [
          {
            "description": "Pin to a run date (default: latest completed batch)",
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Pin to a run date (default: latest completed batch)",
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Risk Summary"
      }
    },
...
    "/api/v1/scenarios/results": {
      "get": {
        "description": "Scenario catalog with per-desk P&L impacts from the latest run that\nexecuted scenarios (EOD runs do; backfill runs don't), worst first.",
        "operationId": "scenarios_results_api_v1_scenarios_results_get",
        "parameters": [
          {
            "description": "Pin to a run date (default: latest completed batch)",
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Pin to a run date (default: latest completed batch)",
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioResults"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scenarios Results"
      }
    },
...
  }
}